refactor: moved custom page renderers to a new module
